l2

+++ deploy.sh
@@ -1,55 +1,34 @@
 #!/usr/bin/sh
 
-function info()
-{
-    echo "Info:" "$@"
-}
-
-function error()
-{
+function fail() {
     echo "Error:" "$@" >&2
     exit 1
 }
 
-function requires()
-{
+function requires() {
     local util=$(command -v $1)
-    if [[ -z "$util" ]]
-    then
-	error "Requires utility" $1
-    else
-	info "Found utility" $util
-    fi
+    [[ -z "$util" ]] && fail "Requires utility" $1
 }
 
-function requires_shell()
-{
+function requires_shell() {
     requires $1
-    if [[ "$SHELL" != "/bin/${1}" ]]
-    then
-	error "Requires shell" $1
-    fi
+    [[ "$SHELL" != "/bin/${1}" ]] && fail "Requires shell" $1
 }
 
-function requires_emacs_version()
-{
+function requires_emacs_version() {
     local version=$(emacs --version | head -n 1 | awk '{print $3}')
-    if [[ $(printf '%s\n' $1 $version | sort -V | head -n1) != $1 ]]
-    then
-	error "Requires at least Emacs version" $1
-    fi
+    [[ $(printf '%s\n' $1 $version | sort -V | head -n1) != $1 ]] &&
+	fail "Requires at least Emacs version" $1
 }
 
-function backup_file()
-{
+function backup_file() {
     [[ ! -e $1 ]] && return
     
     local d=$(dirname $1)
     local b=$(basename $1)
     local n=1
 
-    function f()
-    {
+    function f() {
 	echo "${d}/${b}.${n}.bak"
     }
 
@@ -61,15 +40,21 @@ function backup_file()
     cp -r $1 $(f)
 }
 
-function link_emacs_dir()
-{
+function link_emacs_dir() {
+    local dot_emacs_dir=$(realpath emacs)
+    [[ ! -d $dot_emacs_dir ]] && fail "Missing Emacs dotfiles"
+    
     local emacs_file="$HOME/.emacs"
     local emacs_dir="$HOME/.emacs.d"
+
     backup_file $emacs_file
     backup_file $emacs_dir
+
     [[ -e $emacs_file ]] && rm $emacs_file
-    [[ -e $emacs_dir ]] && rm $emacs_dir
-    ln -sf $(realpath emacs) $emacs_dir
+    [[ -L $emacs_dir ]] && rm $emacs_dir
+    [[ -d $emacs_dir ]] && rm -rf $emacs_dir
+
+    ln -s $dot_emacs_dir $emacs_dir
 }
 
 requires_shell bash
@@ -80,5 +65,5 @@ requires_emacs_version "29.3"
 requires sway
 requires waybar
 
-cd $(dirname $0)
+[[ $0 != $SHELL ]] && cd $(dirname $0)
 link_emacs_dir
